// frontend/src/api/model/githubCom4H1RZooraInternalDomainNotificationAudienceDTOType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType = typeof GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType[keyof typeof GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType];


export const GithubCom4H1RZooraInternalDomainNotificationAudienceDTOType = {
  all: 'all',
  org: 'org',
  class: 'class',
  role: 'role',
  users: 'users',
} as const;